replaced BigDecimal with Double

+++ src/main/java/io/contek/invoker/binancespot/api/common/_AccountBalance.java
@@ -1,12 +1,11 @@
 package io.contek.invoker.binancespot.api.common;
 
 import javax.annotation.concurrent.NotThreadSafe;
-import java.math.BigDecimal;
 
 @NotThreadSafe
 public class _AccountBalance {
 
   public String asset;
-  public BigDecimal free;
-  public BigDecimal locked;
+  public Double free;
+  public Double locked;
 }
