+++ src/main/java/io/contek/invoker/binancespot/api/common/_MarginAsset.java
@@ -1,15 +1,14 @@
 package io.contek.invoker.binancespot.api.common;
 
 import javax.annotation.concurrent.NotThreadSafe;
-import java.math.BigDecimal;
 
 @NotThreadSafe
 public class _MarginAsset {
 
   public String asset;
-  public BigDecimal borrowed;
-  public BigDecimal free;
-  public BigDecimal interest;
-  public BigDecimal locked;
-  public BigDecimal netAsset;
+  public Double borrowed;
+  public Double free;
+  public Double interest;
+  public Double locked;
+  public Double netAsset;
 }
