+++ src/main/java/io/contek/invoker/binancespot/api/common/_NextHourRate.java
@@ -1,11 +1,10 @@
 package io.contek.invoker.binancespot.api.common;
 
 import javax.annotation.concurrent.NotThreadSafe;
-import java.math.BigDecimal;
 
 @NotThreadSafe
 public class _NextHourRate {
 
   public String asset;
-  public BigDecimal nextHourlyInterestRate;
+  public Double nextHourlyInterestRate;
 }
