+++ src/main/java/io/contek/invoker/binancespot/api/common/_Trade.java
@@ -1,7 +1,6 @@
 package io.contek.invoker.binancespot.api.common;
 
 import javax.annotation.concurrent.NotThreadSafe;
-import java.math.BigDecimal;
 
 @NotThreadSafe
 public class _Trade {
@@ -10,10 +9,10 @@ public class _Trade {
   public String id;
   public String orderId;
   public long orderListId;
-  public BigDecimal price;
-  public BigDecimal qty;
-  public BigDecimal quoteQty;
-  public BigDecimal commission;
+  public Double price;
+  public Double qty;
+  public Double quoteQty;
+  public Double commission;
   public String commissionAsset;
   public long time;
   public boolean isBuyer;
