+++ src/main/java/io/contek/invoker/binancespot/api/websocket/user/OrderUpdateChannel.java
@@ -7,7 +7,6 @@ import io.contek.invoker.binancespot.api.websocket.user.constants.UserEventTypeK
 import javax.annotation.concurrent.Immutable;
 import javax.annotation.concurrent.NotThreadSafe;
 import javax.annotation.concurrent.ThreadSafe;
-import java.math.BigDecimal;
 
 @ThreadSafe
 public final class OrderUpdateChannel extends UserWebSocketChannel<OrderUpdateChannel.Data> {
@@ -39,22 +38,22 @@ public final class OrderUpdateChannel extends UserWebSocketChannel<OrderUpdateCh
         public String S; // side
         public String o; // order type
         public String f; // time in force
-        public BigDecimal q; // Order quantity
-        public BigDecimal p; // Order price
-        public BigDecimal P; // stop price
+        public Double q; // Order quantity
+        public Double p; // Order price
+        public Double P; // stop price
         public int d; // Trailing Delta; This is only visible if the order was a trailing stop order.
-        public BigDecimal F; // iceberg qty
+        public Double F; // iceberg qty
         public int g; // order list id
         public String C; // original client order id
         public String x; // Current execution type
         public String X; // current order status
         public String r; // Order reject reason; will be an error code
         public String i; // order id
-        public BigDecimal l; // order last filled quantity
-        public BigDecimal z; // order filled accumulated quantity
-        public BigDecimal L; // Last executed price
+        public Double l; // order last filled quantity
+        public Double z; // order filled accumulated quantity
+        public Double L; // Last executed price
         public String N; // commission asset
-        public BigDecimal n; // commission amount
+        public Double n; // commission amount
         public long T; // order trade time
         public String t; // trade id
         public long I; // ignore
@@ -63,9 +62,9 @@ public final class OrderUpdateChannel extends UserWebSocketChannel<OrderUpdateCh
         public boolean m; // is the trade the maker side?
         public boolean M; // ignore
         public long O; // order creation time
-        public BigDecimal Z; // Cumulative quote asset transacted quantity
-        public BigDecimal Y; // Last quote asset transacted quantity (i.e. lastPrice * lastQty)
-        public BigDecimal Q; // Quote Order Qty
+        public Double Z; // Cumulative quote asset transacted quantity
+        public Double Y; // Last quote asset transacted quantity (i.e. lastPrice * lastQty)
+        public Double Q; // Quote Order Qty
         public long j; // strategy id. This is only visible if the strategyId parameter was provided upon order placement
         public long J; // strategy type. This is only visible if the strategy type parameter was provided upon order placement
         public long traceNano = System.nanoTime();
